server/requirements.txt
Flask
Flask-Cors
pydantic
Werkzeug
PyJWT
easyocr
openpyxl
reportlab
Pillow
torch
transformers
waitress